fix: UI and build error for explorer plugin
