+++ .github/workflows/on-main-push.yaml
@@ -13,7 +13,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        package: [ blueprint, default-pdf, default-preview, job, mermaid, yaml-view, dm-core, tabs, header, form]
+        package: [ blueprint, default-pdf, default-preview, job, mermaid, yaml-view, dm-core, tabs, header, form, explorer]
     steps:
       - uses: actions/checkout@v2
