+++ example/app/data/DemoDataSource/DemoPackage/application.json
@@ -1,4 +1,5 @@
 {
+  "_id": "03bf685b-edb6-40e4-8c67-62b13fefecaa",
   "name": "exampleApp",
   "type": "CORE:Application",
   "visibleDataSources": ["system", "DemoDataSource", "WorkflowDS"]
