+++ packages/explorer/package.json
@@ -53,12 +53,13 @@
   },
   "peerDependencies": {
     "react": ">=17.0.2",
-    "@development-framework/dm-core": ">=1.0.25",
+    "@development-framework/dm-core": ">=1.0.33",
     "styled-components": ">=5.3.6"
   },
   "scripts": {
     "prebuild": "rm -rf dist && yarn install",
-    "build": "tsc",
+    "copy-css-file": "cp src/components/context-menu/react-contextmenu.css  dist/components/context-menu/react-contextmenu.css",
+    "build": "tsc && yarn copy-css-file",
     "postbuild": "cp package.json ./dist/ && rm -rf node_modules"
   }
 }
