+++ packages/explorer/src/components/context-menu/utils/contextMenuActions.ts
@@ -1,0 +1,88 @@
+import {
+  ErrorResponse,
+  TreeNode,
+  DmssAPI,
+} from '@development-framework/dm-core'
+import { AxiosError } from 'axios'
+// @ts-ignore
+import { NotificationManager } from 'react-notifications'
+
+export const DeleteAction = async (
+  node: TreeNode,
+  dmssAPI: DmssAPI,
+  setLoading: (isLoading: boolean) => void
+) => {
+  setLoading(true)
+  await dmssAPI
+    .documentRemoveByPath({
+      dataSourceId: node.dataSource,
+      directory: node.pathFromRootPackage(),
+    })
+    .then(() => {
+      node.remove()
+      NotificationManager.success('Deleted')
+    })
+    .catch((error: AxiosError<ErrorResponse>) => {
+      console.error(error)
+      NotificationManager.error(
+        error.response?.data.message,
+        'Failed to delete'
+      )
+    })
+    .finally(() => setLoading(false))
+}
+
+export const NewFolderAction = (
+  node: TreeNode,
+  folderName: string,
+  dmssAPI: DmssAPI,
+  setLoading: (isLoading: boolean) => void
+) => {
+  const newFolder = {
+    name: folderName,
+    type: 'dmss://system/SIMOS/Package',
+    isRoot: false,
+    content: [],
+  }
+  const ref = `${node.nodeId}.content`
+  dmssAPI
+    .documentAdd({
+      absoluteRef: ref,
+      body: newFolder,
+      updateUncontained: true,
+    })
+    .then(() => node.expand())
+    .catch((error: AxiosError<ErrorResponse>) => {
+      console.error(error)
+      NotificationManager.error(error.response?.data.message)
+    })
+}
+export const NewRootPackageAction = (
+  node: TreeNode,
+  packageName: string,
+  dmssAPI: DmssAPI,
+  setLoading: (isLoading: boolean) => void
+) => {
+  const newPackage = {
+    name: packageName,
+    type: 'dmss://system/SIMOS/Package',
+    isRoot: true,
+    content: [],
+  }
+  const ref: string = node.dataSource
+  dmssAPI
+    .documentAdd({
+      absoluteRef: ref,
+      body: newPackage,
+      updateUncontained: true,
+    })
+    .then(() => {
+      node.expand()
+    })
+    .catch((error: AxiosError<ErrorResponse>) => {
+      NotificationManager.error(
+        JSON.stringify(error.response?.data.message),
+        'Failed to create new root package'
+      )
+    })
+}
