+++ packages/explorer/src/components/context-menu/utils/styledComponents.ts
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+
+export const edsButtonStyleConfig = {
+  marginTop: '40px',
+  width: '200px',
+  alignSelf: 'center',
+}
+
+export const DialogContent = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+`
