+++ packages/explorer/src/index.tsx
@@ -1,5 +1,5 @@
-import {EPluginType, TPlugin} from '@development-framework/dm-core'
-import Editor from "./Editor";
+import { EPluginType, TPlugin } from '@development-framework/dm-core'
+import Editor from './Editor'
 
 export const plugins: TPlugin[] = [
   {
