+++ packages/explorer/tsconfig.json
@@ -2,5 +2,7 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "outDir": "dist"
-  }
+  },
+  "include": ["src"],
+  "exclude": ["node_modules", "src/**/*.spec.ts", "dist"]
 }
